refactor(navbar-type): Refactored NavInterface and made LinkData type, updating usage as necessary
