+++ client/src/components/Navbar/Sidebar.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Link, useNavigate } from "react-router-dom"
 
-import { NavbarProps } from "../../types/interfaces/NavInterface"
+import { NavbarProps } from "../../types/interfaces/Navbar/NavInterface"
 
 import moNewsLogo from "/assets/mo_news_logo_white.png"
