--- client/src/types/interfaces/NavInterface.tsx
@@ -1,9 +1,0 @@
-interface NavbarProps {
-    links: { 
-        name: string;
-         url: string
-         }[];
-}
-
-
-export type {NavbarProps}
