+++ client/src/types/interfaces/Navbar/NavInterface.tsx
@@ -1,0 +1,8 @@
+import { LinkData } from "./LinkData"
+
+interface NavbarProps {
+    links: LinkData[]
+}
+
+
+export type {NavbarProps}
